services/connections/requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
kafka-python